# Initialize markdown only once