This is a small performance improvement. But it should be somewhat significant for large collections with a lot of markdown.