@Glandos 2018-06-17
Notice that this is about performance for no-op builds. Otherwise, image resizing and video compression is far more important for the CPU.