@Glandos 2018-07-05
OK, here are more figures. For brevity, I skip output from `sigal`. I manually applied the [patch for python-markdown](https://github.com/Python-Markdown/markdown/pull/672). All sigal tests are still OK.

### `master` with `photoswipe` theme
```
 Performance counter stats for './sigal_py3_venv/bin/sigal build photos/ _build_prod' (10 runs):

       5015,872216      task-clock (msec)         #    0,990 CPUs utilized            ( +-  5,22% )
               676      context-switches          #    0,135 K/sec                    ( +-  7,62% )
                 9      cpu-migrations            #    0,002 K/sec                    ( +- 15,99% )
            13 899      page-faults               #    0,003 M/sec                    ( +-  0,83% )
    12 944 448 854      cycles                    #    2,581 GHz                      ( +-  0,17% )  (49,95%)
    11 274 264 137      instructions              #    0,87  insn per cycle           ( +-  0,10% )  (74,98%)
     2 354 613 689      branches                  #  469,433 M/sec                    ( +-  0,13% )  (75,04%)
       115 908 249      branch-misses             #    4,92% of all branches          ( +-  0,17% )  (75,01%)

       5,067768353 seconds time elapsed                                          ( +-  5,26% )
```

### This patch with `photoswipe` theme
```
 Performance counter stats for './sigal_py3_venv/bin/sigal build photos/ _build_prod' (10 runs):

       4794,963802      task-clock (msec)         #    0,963 CPUs utilized            ( +-  3,02% )
               922      context-switches          #    0,192 K/sec                    ( +- 13,65% )
                13      cpu-migrations            #    0,003 K/sec                    ( +- 26,46% )
            14 410      page-faults               #    0,003 M/sec                    ( +-  0,82% )
    12 704 923 530      cycles                    #    2,650 GHz                      ( +-  0,84% )  (50,01%)
    10 998 472 366      instructions              #    0,87  insn per cycle           ( +-  0,16% )  (75,02%)
     2 297 597 621      branches                  #  479,169 M/sec                    ( +-  0,17% )  (75,04%)
       112 298 936      branch-misses             #    4,89% of all branches          ( +-  0,38% )  (74,95%)

       4,977598037 seconds time elapsed                                          ( +-  3,55% )
```
This is 1.77% faster, so it is insignificant.

### `master` branch with a custom theme
This is a theme I am currently working on. It's a derivative of `photoswipe`
```
 Performance counter stats for './sigal_py3_venv/bin/sigal build photos/ _build_prod' (10 runs):

       2219,313680      task-clock (msec)         #    0,980 CPUs utilized            ( +-  4,67% )
               428      context-switches          #    0,193 K/sec                    ( +-  5,91% )
                 8      cpu-migrations            #    0,004 K/sec                    ( +- 10,91% )
            12 130      page-faults               #    0,005 M/sec                    ( +-  0,03% )
     5 711 295 676      cycles                    #    2,573 GHz                      ( +-  0,39% )  (50,02%)
     4 948 863 961      instructions              #    0,87  insn per cycle           ( +-  0,21% )  (75,08%)
     1 044 639 102      branches                  #  470,704 M/sec                    ( +-  0,14% )  (75,13%)
        50 406 356      branch-misses             #    4,83% of all branches          ( +-  0,25% )  (74,85%)

       2,263710201 seconds time elapsed                                          ( +-  4,44% )
```
Here I'm starting to have a doubt.

### This patch with my custom theme
```
 Performance counter stats for './sigal_py3_venv/bin/sigal build photos/ _build_prod' (10 runs):

       2064,062561      task-clock (msec)         #    0,970 CPUs utilized            ( +-  4,44% )
               434      context-switches          #    0,210 K/sec                    ( +-  5,50% )
                 9      cpu-migrations            #    0,004 K/sec                    ( +- 13,92% )
            12 767      page-faults               #    0,006 M/sec                    ( +-  0,72% )
     5 379 681 810      cycles                    #    2,606 GHz                      ( +-  0,35% )  (49,92%)
     4 665 004 022      instructions              #    0,87  insn per cycle           ( +-  0,07% )  (74,98%)
       985 776 218      branches                  #  477,590 M/sec                    ( +-  0,11% )  (75,06%)
        47 268 135      branch-misses             #    4,80% of all branches          ( +-  0,16% )  (75,02%)

       2,126881752 seconds time elapsed                                          ( +-  4,42% )
```
And we still have a 6% improvement. Not as significant as before, but still.

### Conclusion
My custom theme seems to spend way less time to write files. I didn't have the time to inspect why for now. But only with it the parsing of Markdown is significant, otherwise, the time spent when writing files is much higher than parsing.

No doubt that this patch improve the markdown parsing. But it seems to be insignificant for a lot of cases. I'd like to see it merged, but I could understand that it will probably be invisible for everyone.